+++ src/main/java/net/dv8tion/jda/api/exceptions/RateLimitedException.java
@@ -16,7 +16,7 @@
 
 package net.dv8tion.jda.api.exceptions;
 
-import net.dv8tion.jda.internal.requests.Route;
+import net.dv8tion.jda.api.requests.Route;
 
 /**
  * Indicates that we received a {@code 429: Too Many Requests} response
