+++ src/main/java/net/dv8tion/jda/api/requests/RestFuture.java
@@ -18,7 +18,6 @@ package net.dv8tion.jda.api.requests;
 
 import net.dv8tion.jda.internal.JDAImpl;
 import net.dv8tion.jda.internal.requests.RestActionImpl;
-import net.dv8tion.jda.internal.requests.Route;
 import okhttp3.RequestBody;
 import org.apache.commons.collections4.map.CaseInsensitiveMap;
