+++ src/main/java/net/dv8tion/jda/internal/entities/UserImpl.java
@@ -25,7 +25,7 @@ import net.dv8tion.jda.api.utils.data.DataObject;
 import net.dv8tion.jda.internal.JDAImpl;
 import net.dv8tion.jda.internal.requests.EmptyRestAction;
 import net.dv8tion.jda.internal.requests.RestActionImpl;
-import net.dv8tion.jda.internal.requests.Route;
+import net.dv8tion.jda.api.requests.Route;
 
 import javax.annotation.Nonnull;
 import java.util.FormattableFlags;
