+++ src/main/java/net/dv8tion/jda/internal/managers/EmoteManagerImpl.java
@@ -25,7 +25,7 @@ import net.dv8tion.jda.api.managers.EmoteManager;
 import net.dv8tion.jda.api.utils.data.DataArray;
 import net.dv8tion.jda.api.utils.data.DataObject;
 import net.dv8tion.jda.internal.entities.EmoteImpl;
-import net.dv8tion.jda.internal.requests.Route;
+import net.dv8tion.jda.api.requests.Route;
 import net.dv8tion.jda.internal.utils.Checks;
 import okhttp3.RequestBody;
