+++ src/main/java/net/dv8tion/jda/internal/managers/WebhookManagerImpl.java
@@ -23,7 +23,7 @@ import net.dv8tion.jda.api.entities.Webhook;
 import net.dv8tion.jda.api.exceptions.InsufficientPermissionException;
 import net.dv8tion.jda.api.managers.WebhookManager;
 import net.dv8tion.jda.api.utils.data.DataObject;
-import net.dv8tion.jda.internal.requests.Route;
+import net.dv8tion.jda.api.requests.Route;
 import net.dv8tion.jda.internal.utils.Checks;
 import okhttp3.RequestBody;
