+++ src/main/java/net/dv8tion/jda/internal/requests/ratelimit/ClientRateLimiter.java
@@ -22,7 +22,7 @@ import net.dv8tion.jda.api.utils.data.DataObject;
 import net.dv8tion.jda.internal.JDAImpl;
 import net.dv8tion.jda.internal.requests.RateLimiter;
 import net.dv8tion.jda.internal.requests.Requester;
-import net.dv8tion.jda.internal.requests.Route;
+import net.dv8tion.jda.api.requests.Route;
 import net.dv8tion.jda.internal.utils.IOUtil;
 
 import java.io.IOException;
