+++ src/main/java/net/dv8tion/jda/internal/requests/restaction/order/OrderActionImpl.java
@@ -19,7 +19,7 @@ package net.dv8tion.jda.internal.requests.restaction.order;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.requests.restaction.order.OrderAction;
 import net.dv8tion.jda.internal.requests.RestActionImpl;
-import net.dv8tion.jda.internal.requests.Route;
+import net.dv8tion.jda.api.requests.Route;
 import net.dv8tion.jda.internal.utils.Checks;
 
 import javax.annotation.Nonnull;
@@ -44,7 +44,7 @@ public abstract class OrderActionImpl<T, M extends OrderAction<T, M>>
      *        JDA instance which is associated with the entities contained
      *        in the order list
      * @param route
-     *        The {@link net.dv8tion.jda.internal.requests.Route.CompiledRoute CompiledRoute}
+     *        The {@link Route.CompiledRoute CompiledRoute}
      *        which is provided to the {@link RestActionImpl#RestActionImpl(JDA, Route.CompiledRoute, okhttp3.RequestBody) RestAction Constructor}
      */
     public OrderActionImpl(JDA api, Route.CompiledRoute route)
@@ -61,7 +61,7 @@ public abstract class OrderActionImpl<T, M extends OrderAction<T, M>>
      * @param ascendingOrder
      *        Whether or not the order of items should be ascending
      * @param route
-     *        The {@link net.dv8tion.jda.internal.requests.Route.CompiledRoute CompiledRoute}
+     *        The {@link Route.CompiledRoute CompiledRoute}
      *        which is provided to the {@link RestActionImpl#RestActionImpl(JDA, Route.CompiledRoute, okhttp3.RequestBody) RestAction Constructor}
      */
     public OrderActionImpl(JDA api, boolean ascendingOrder, Route.CompiledRoute route)
